% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{example_proportions}
\alias{example_proportions}
\title{Example Cell Proportions Dataset}
\format{
A data frame with 2 rows and 3 variables:
\describe{
  \item{sample_id}{Character, the ID of the sample}
  \item{cell_type}{Character, the type of cell - L2 annotation}
  \item{proportion}{Numeric, the proportion of that cell type}
}
}
\usage{
data(example_proportions)
}
\description{
A small dataset containing example cell type proportions for testing.
}
\examples{
data(example_proportions)
head(example_proportions)
}
\keyword{datasets}
