% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{get_default_cache_dir}
\alias{get_default_cache_dir}
\title{Returns the default cache directory with a version number}
\usage{
get_default_cache_dir()
}
\value{
A length one character vector.
}
\description{
Returns the default cache directory with a version number
}
