% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{get_sub_formula}
\alias{get_sub_formula}
\title{Extract sub-formula from full formula}
\usage{
get_sub_formula(full_formula, factor_names)
}
\arguments{
\item{full_formula}{A full formula object.}

\item{factor_names}{A vector of factor names to retain.}
}
\value{
A sub-formula as a string.
}
\description{
This function extracts a sub-formula that contains only the specified factor names.
}
