% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{posterior_test}
\alias{posterior_test}
\title{posterior_test: Perform posterior predictive analysis}
\usage{
posterior_test(
  proportions = NULL,
  sex = NULL,
  age_bin = NULL,
  disease_groups = NULL,
  ethnicity_groups = NULL,
  assay_groups = NULL,
  tissue_groups = NULL,
  path_to_model =
    "~/R/SAiGENCI/immune_map/pseudobulk_sample_is_immune/estimates_age_bins___L2.rds"
)
}
\arguments{
\item{proportions}{A data frame of observed cell type proportions.}

\item{sex}{A string indicating sex ("male", "female", or "unknown").}

\item{age_bin}{A string for age bin category.}

\item{disease_groups}{A string indicating disease group.}

\item{ethnicity_groups}{A string indicating ethnicity group.}

\item{assay_groups}{A string indicating assay group.}

\item{tissue_groups}{A string indicating tissue group.}

\item{path_to_model}{Path to the `.rds` posterior model file.}
}
\value{
A list with a result table and a density plot.
}
\description{
This function performs posterior predictive sampling and returns
a table and a density plot.
}
