% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{expr_predict}
\alias{expr_predict}
\title{expr_predict: Predict gene expression using brms model}
\usage{
expr_predict(
  cell_type,
  gene_ensg,
  age_decade = NA,
  sex = NA,
  disease_groups = NA,
  ethnicity_groups = NA,
  assay_groups = NA,
  tissue_groups = NA
)
}
\arguments{
\item{cell_type}{A string indicating the cell type.}

\item{gene_ensg}{A string indicating the gene ENSEMBL ID (e.g., "ENSG00000000419").}

\item{age_decade}{A string or NA for age decade (default: NA).}

\item{sex}{A string or NA for sex (default: NA).}

\item{disease_groups}{A string or NA for disease groups (default: NA).}

\item{ethnicity_groups}{A string or NA for ethnicity groups (default: NA).}

\item{assay_groups}{A string or NA for assay groups (default: NA).}

\item{tissue_groups}{A string or NA for tissue groups (default: NA).}
}
\value{
A list containing predictions (pred), summary statistics (mean, median, peak location), and a density plot.
}
\description{
This function downloads a brms model file for a specific cell type and gene,
performs prediction based on provided metadata, and returns predictions with
summary statistics and a density plot.
}
