% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{get_model_ready}
\alias{get_model_ready}
\title{Get the file path for the pre-trained model}
\usage{
get_model_ready(cache_directory = get_default_cache_dir(), use_cache = TRUE)
}
\arguments{
\item{cache_directory}{Character. Directory to store the cached file. Defaults to the standard cache directory.}

\item{use_cache}{Logical. If TRUE, uses the cached version if available.}
}
\value{
Character. The local file path of the model.
}
\description{
Checks if the pre-trained model file is cached locally. If not, downloads it from the cloud.
}
